## client - renamed page title to Dogalogue - renamed `FeedPosts.jsx` to `Posts.jsx` - moved individual post to its own component `Post.jsx` - formatted string in footer of `SinglePostModal.jsx` - cleanded up imports in `DogFeed.jsx` - refactored to use `Posts` instead of `FeedPosts` in `DogFeed.jsx` - added post fetching to `PetPage.jsx` - show pet's posts on pet page - add `QUERY_PET_POSTS` query
## server
- remove unused console.log from pet resolver
- added `petPosts` resolver
- add owner user to pet when registering
- added `petPosts` to `Query` type in `typeDefs.js`
