+++ client/src/pages/DogFeed.jsx
@@ -1,16 +1,9 @@
-import FeedPosts from "../components/FeedPosts";
-import FeedPostInput from "../components/FeedPostInput";
+import Posts from "../components/Posts";
 import Auth from "../utils/auth";
-
-import Modal from "react-bootstrap/Modal";
-import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 
 import { useState } from "react";
-import auth from "../utils/auth";
-import { useMutation, useQuery } from "@apollo/client";
-import { CREATE_POST } from "../utils/mutations";
-import { uploadImage } from "../utils/images";
+import { useQuery } from "@apollo/client";
 import NewPostModal from "../components/NewPostModal";
 import { QUERY_POSTS } from "../utils/queries";
 
@@ -35,14 +28,14 @@ const DogFeed = () => {
 
   return (
     <>
-      <div className="feedPage full-withradius border">
+      <div className="feedPage full-withradius">
         {/* <pre>{JSON.stringify(Auth.getProfile())}</pre> */}
         <header className="d-flex w-100 justify-content-between">
           <h1 className="feed-title">Posts</h1>
           <Button onClick={() => setShowNewPostModal(true)}>New Post</Button>
         </header>
 
-        <FeedPosts posts={data.posts} />
+        <Posts posts={data.posts} />
       </div>
 
       {/* new post modal */}
